refactor: move handleSeerCommandError to shared utils
Move the error handling helper from plan.ts to utils.ts so both
explain.ts and plan.ts use the same function. This ensures outcome
classification stays consistent if the logic is updated.
